(Improve order flow) Remove submitters (#4062) (minor)
### Changed

- The OrderSubmitterRegistry is a StoringServiceRegistry
- The OrderSubmitters are removed
- The OrdersAPI needs no Store in its init
- The OrdersAPI needs a StoringServiceRegistry instead of a SubmitterRegistry
- The OrdersAPI needs an OrderValidationService
