+++ cg/services/orders/order_service/order_service.py
@@ -1,13 +1,10 @@
 from cg.server.dto.orders.orders_request import OrdersRequest
-from cg.server.dto.orders.orders_response import Order as OrderResponse, Order
-from cg.server.dto.orders.orders_response import OrdersResponse
+from cg.server.dto.orders.orders_response import Order, OrdersResponse
 from cg.services.orders.order_service.models import OrderQueryParams
 from cg.services.orders.order_summary_service.dto.order_summary import OrderSummary
-from cg.services.orders.order_summary_service.order_summary_service import (
-    OrderSummaryService,
-)
-from cg.store.store import Store
+from cg.services.orders.order_summary_service.order_summary_service import OrderSummaryService
 from cg.store.models import Order as DatabaseOrder
+from cg.store.store import Store
 
 
 class OrderService:
@@ -15,8 +12,8 @@ class OrderService:
         self.store = store
         self.summary_service = status_service
 
-    def get_order(self, order_id: int) -> OrderResponse:
-        order: Order = self.store.get_order_by_id(order_id)
+    def get_order(self, order_id: int) -> Order:
+        order: DatabaseOrder = self.store.get_order_by_id(order_id)
         summary: OrderSummary = self.summary_service.get_summary(order_id)
         return self._create_order_response(order=order, summary=summary)
 
@@ -29,13 +26,13 @@ class OrderService:
         summaries: list[OrderSummary] = self.summary_service.get_summaries(order_ids)
         return self._create_orders_response(orders=orders, summaries=summaries, total=total_count)
 
-    def set_open(self, order_id: int, open: bool) -> OrderResponse:
-        order: Order = self.store.update_order_status(order_id=order_id, open=open)
+    def set_open(self, order_id: int, open: bool) -> Order:
+        order: DatabaseOrder = self.store.update_order_status(order_id=order_id, open=open)
         return self._create_order_response(order)
 
     def update_is_open(self, order_id: int, delivered_analyses: int) -> None:
         """Update the is_open parameter of an order based on the number of delivered analyses."""
-        order: Order = self.store.get_order_by_id(order_id)
+        order: DatabaseOrder = self.store.get_order_by_id(order_id)
         case_count: int = len(order.cases)
         if self._is_order_closed(case_count=case_count, delivered_analyses=delivered_analyses):
             self.set_open(order_id=order_id, open=False)
