+++ cg/services/orders/store_order_services/store_order_service.py
@@ -4,8 +4,8 @@ import logging
 from abc import ABC, abstractmethod
 
 from cg.models.orders.order import OrderIn
+from cg.services.order_validation_service.models.order import Order
 from cg.services.order_validation_service.models.sample import Sample
-from cg.services.order_validation_service.order_validation_service import OrderValidationService
 from cg.services.orders.order_lims_service.order_lims_service import OrderLimsService
 from cg.store.store import Store
 
@@ -29,7 +29,7 @@ class StoreOrderService(ABC):
         self.lims = lims_service
 
     @abstractmethod
-    def store_order(self, order_in: OrderIn):
+    def store_order(self, order: Order):
         pass
 
     @staticmethod
@@ -40,18 +40,3 @@ class StoreOrderService(ABC):
             internal_id = lims_map[sample.name]
             LOG.info(f"{sample.name} -> {internal_id}: connect sample to LIMS")
             sample._generated_lims_id = internal_id
-
-
-class OrderSubmitter(ABC):
-    @abstractmethod
-    def __init__(
-        self,
-        validate_order_service: OrderValidationService,
-        store_order_service: StoreOrderService,
-    ):
-        self.order_validation_service = validate_order_service
-        self.order_store_service = store_order_service
-
-    @abstractmethod
-    def submit_order(self, order_in: OrderIn) -> dict:
-        pass
