--- cg/services/orders/submitters/metagenome_order_submitter.py
@@ -1,25 +1,0 @@
-from cg.services.order_validation_service.workflows.metagenome.models.order import MetagenomeOrder
-from cg.services.order_validation_service.workflows.taxprofiler.models.order import TaxprofilerOrder
-from cg.services.orders.store_order_services.store_metagenome_order import (
-    StoreMetagenomeOrderService,
-)
-from cg.services.orders.submitters.order_submitter import OrderSubmitter
-from cg.services.orders.validate_order_services.validate_metagenome_order import (
-    ValidateMetagenomeOrderService,
-)
-
-
-class MetagenomeOrderSubmitter(OrderSubmitter):
-    """Class for submitting metagenome and taxprofiler orders."""
-
-    def __init__(
-        self,
-        order_validation_service: ValidateMetagenomeOrderService,
-        order_store_service: StoreMetagenomeOrderService,
-    ):
-        self.order_validation_service = order_validation_service
-        self.order_store_service = order_store_service
-
-    def submit_order(self, order: MetagenomeOrder | TaxprofilerOrder) -> dict:
-        """Submit a metagenome order."""
-        return self.order_store_service.store_order(order)
