+++ cg/services/orders/validate_order_services/validate_fastq_order.py
@@ -1,5 +1,5 @@
 from cg.models.orders.order import OrderIn
-from cg.services.orders.submitters.order_submitter import ValidateOrderService
+from cg.services.orders.store_order_services.store_order_service import ValidateOrderService
 from cg.store.store import Store
