+++ cg/services/orders/validate_order_services/validate_pacbio_order.py
@@ -1,7 +1,7 @@
 from cg.exc import OrderError
 from cg.models.orders.order import OrderIn
 from cg.models.orders.samples import PacBioSample
-from cg.services.orders.submitters.order_submitter import ValidateOrderService
+from cg.services.orders.store_order_services.store_order_service import ValidateOrderService
 from cg.store.models import ApplicationVersion, Customer
 from cg.store.store import Store
