+++ tests/fixture_plugins/orders_fixtures/store_service_fixtures.py
@@ -1,19 +1,27 @@
 import pytest
 
 from cg.services.orders.order_lims_service.order_lims_service import OrderLimsService
-from cg.services.orders.store_order_services.store_case_order import StoreCaseOrderService
-from cg.services.orders.store_order_services.store_fastq_order_service import StoreFastqOrderService
-from cg.services.orders.store_order_services.store_metagenome_order import (
+from cg.services.orders.store_order_services.implementations.store_case_order import (
+    StoreCaseOrderService,
+)
+from cg.services.orders.store_order_services.implementations.store_fastq_order_service import (
+    StoreFastqOrderService,
+)
+from cg.services.orders.store_order_services.implementations.store_metagenome_order import (
     StoreMetagenomeOrderService,
 )
-from cg.services.orders.store_order_services.store_microbial_fastq_order_service import (
+from cg.services.orders.store_order_services.implementations.store_microbial_fastq_order_service import (
     StoreMicrobialFastqOrderService,
 )
-from cg.services.orders.store_order_services.store_microbial_order import StoreMicrobialOrderService
-from cg.services.orders.store_order_services.store_pacbio_order_service import (
+from cg.services.orders.store_order_services.implementations.store_microbial_order import (
+    StoreMicrobialOrderService,
+)
+from cg.services.orders.store_order_services.implementations.store_pacbio_order_service import (
     StorePacBioOrderService,
 )
-from cg.services.orders.store_order_services.store_pool_order import StorePoolOrderService
+from cg.services.orders.store_order_services.implementations.store_pool_order import (
+    StorePoolOrderService,
+)
 from cg.store.store import Store
 from tests.mocks.limsmock import MockLimsAPI
 from tests.store_helpers import StoreHelpers
