+++ tests/services/orders/order_store_service/test_microbial_fastq_order_store_service.py
@@ -2,7 +2,7 @@ from cg.constants import DataDelivery
 from cg.services.order_validation_service.workflows.microbial_fastq.models.order import (
     MicrobialFastqOrder,
 )
-from cg.services.orders.store_order_services.store_microbial_fastq_order_service import (
+from cg.services.orders.store_order_services.implementations.store_microbial_fastq_order_service import (
     StoreMicrobialFastqOrderService,
 )
 from cg.store.models import Case, Sample
